tool_operate: cast upload_flags properly

+++ src/tool_operate.c
@@ -1758,7 +1758,7 @@ static CURLcode config2setopts(struct GlobalConfig *global,
   }
   /* new in 8.13.0 */
   if(config->upload_flags)
-    my_setopt(curl, CURLOPT_UPLOAD_FLAGS, config->upload_flags);
+    my_setopt(curl, CURLOPT_UPLOAD_FLAGS, (long)config->upload_flags);
   return result;
 }
